A mocks folder was created, and functions for the useTranslations hook were implemented. The Help and Search components were modified to incorporate translations for their messages
