+++ src/app/[locale]/style/components/search.css
@@ -9,7 +9,7 @@
     transition: opacity 0.5s ease, transform 0.5s ease;
     animation: fadeIn 0.5s forwards;
     position: relative;
-    z-index: 999;
+    z-index: 100;
 }
 
 .p7 .search {
