+++ src/app/[locale]/ticket/page.jsx
@@ -8,29 +8,19 @@ import styles from './ticket.module.css';
 import Input from "@/components/input";
 import { sendContactForm } from "@/library/send";
 import Editor from "../create-vacancy/editor";
+import { getOptions, getRoles } from "@/app/mocks/helps";
+import Search from "@/components/header/search";
+import { useTranslations } from 'next-intl';
 
 export default function Ticket() {
+    const t = useTranslations();
     const [role, setRole] = useState('notSelect');
     const [option, setOption] = useState('notSelect');
-    const [counter, setCounter] = useState(100)
-    const [topic, setTopic] = useState('')
-    const [name, setName] = useState("")
-    const [email, setEmail] = useState("")
-    const [description, setDescription] = useState("<h1>Сообщения</h1> <ul><li></li><li></li></ul>")
-
-    const roles = [
-        { value: 'notSelect', label: 'Не выбрано' },
-        { value: 'employer', label: 'Соискатель' },
-        { value: 'employee', label: 'Работодатель' }
-    ];
-
-    
-    const options = [
-        { value: 'notSelect', label: 'Не выбрано' },
-        { value: 'question', label: 'Задать вопрос/рассказать о проблеме' },
-        { value: 'support', label: 'Оформить жалобу' },
-        { value: 'thanks', label: 'Выразить благодарность' }
-    ];
+    const [counter, setCounter] = useState(100);
+    const [topic, setTopic] = useState('');
+    const [name, setName] = useState('');
+    const [email, setEmail] = useState('');
+    const [description, setDescription] = useState('<h1>Сообщения</h1> <ul><li></li><li></li></ul>');
 
     const handleSubmit = async () => {
         const data = {
@@ -44,42 +34,40 @@ export default function Ticket() {
         console.log(data);
         await sendContactForm(data);
     };
-    
 
-    useEffect(() => {
-        setCounter(100 - topic.length)
-    }, [topic])
-    console.log(description)
+    useEffect(() => { setCounter(100 - topic.length) }, [topic]);
+
     return (
         <div className="wrapper">
-        <Header />
-        <main>
-            <span className={styles.write}>Написать письмо</span>
-            <div className="container midContainer">
-                <label className={styles.role} htmlFor="role">Я обращаюсь как</label>
-                <CustomSelect
-                    options={roles}
-                    value={role}
-                    onChange={setRole}
-                    placeholder="Не выбрано"
-                />
-                <div className="mtb4"></div>
-                <label className={styles.role} htmlFor="role">Выберите нужную опцию </label>
-                <CustomSelect
-                    options={options}
-                    value={option}
-                    onChange={setOption}
-                    placeholder="Не выбрано"
-                />
-                <div className="mtb4"></div>
-                <Input type='text' counter={counter} label='Тема' onChange={(e) => setTopic(e.target.value)}/>
-                <Editor description={description} setDescription={setDescription}/>
-                <Input placeholder='Мади Ерасылович' label="Имя" type='text' onChange={(e) => setName(e.target.value)}/>
-                <Input placeholder='example@com' label="Почта" type='text' onChange={(e) => setEmail(e.target.value)}/>
-                <button onClick={handleSubmit} className="button button-black">Продолжить</button>
-            </div>
-        </main>
-        <Footer />
+            <Header />
+            <main>
+                <span className={styles.write}>{t('ticket.writeLetter')}</span>
+                <div className="container midContainer">
+                    <Search />
+                    <label className={styles.role} htmlFor="role">{t('ticket.contactAs')}</label>
+                    <CustomSelect
+                        options={getRoles(t)}
+                        value={role}
+                        onChange={setRole}
+                        placeholder={t('ticket.notSelected')}
+                    />
+                    <div className="mtb4"></div>
+                    <label className={styles.role} htmlFor="role">{t('ticket.selectOption')}</label>
+                    <CustomSelect
+                        options={getOptions(t)}
+                        value={option}
+                        onChange={setOption}
+                        placeholder={t('ticket.notSelected')}
+                    />
+                    <div className="mtb4"></div>
+                    <Input type='text' counter={counter} label={t('ticket.topic')} onChange={(e) => setTopic(e.target.value)} />
+                    <Editor description={description} setDescription={setDescription} />
+                    <Input placeholder={t('ticket.placeholderName')} label={t('ticket.name')} type='text' onChange={(e) => setName(e.target.value)} />
+                    <Input placeholder={t('ticket.placeholderEmail')} label={t('ticket.email')} type='text' onChange={(e) => setEmail(e.target.value)} />
+                    <button onClick={handleSubmit} className="button button-black">{t('ticket.continue')}</button>
+                </div>
+            </main>
+            <Footer />
         </div>
     );
 }
