+++ src/app/[locale]/ticket/ticket.module.css
@@ -13,5 +13,6 @@
 
 .role::after {
     content: "*";
-    color: red;
+    color: rgb(241, 61, 61);
+    padding-left: 5px;
 }
