+++ src/app/mocks/helps.jsx
@@ -1,0 +1,12 @@
+export const getRoles = (t) => [
+    { value: 'notSelect', label: t('roles.notSelect') },
+    { value: 'employer', label: t('roles.employer') },
+    { value: 'employee', label: t('roles.employee') }
+];
+
+export const getOptions = (t) => [
+    { value: 'notSelect', label: t('options.notSelect') },
+    { value: 'question', label: t('options.question') },
+    { value: 'support', label: t('options.support') },
+    { value: 'thanks', label: t('options.thanks') }
+];
