+++ src/app/mocks/suggestion.jsx
@@ -1,0 +1,251 @@
+export const suggestion = [
+    "Программист",
+    "Веб-разработчик",
+    "Менеджер проектов",
+    "Системный администратор",
+    "Графический дизайнер",
+    "Маркетолог",
+    "Контент-менеджер",
+    "Копирайтер",
+    "Специалист по продажам",
+    "Финансовый аналитик",
+    "Бухгалтер",
+    "Специалист по работе с клиентами",
+    "Юрист",
+    "Секретарь",
+    "Переводчик",
+    "SEO-специалист",
+    "Видеомонтажёр",
+    "Инженер-программист",
+    "Врач",
+    "Медсестра",
+    "Учитель",
+    "Педагог",
+    "Архитектор",
+    "Химик",
+    "Физик",
+    "Биолог",
+    "Специалист по кадрам",
+    "Логист",
+    "Экономист",
+    "Продавец-консультант",
+    "Массажист",
+    "Стоматолог",
+    "Психолог",
+    "Специалист по контекстной рекламе",
+    "Администратор сайта",
+    "Интернет-маркетолог",
+    "Аналитик данных",
+    "Аудитор",
+    "Тестировщик",
+    "Водитель",
+    "Электрик",
+    "Менеджер по персоналу",
+    "Директор",
+    "Тренер",
+    "Дизайнер интерьеров",
+    "Рекрутер",
+    "Менеджер по закупкам",
+    "Курьер",
+    "Агроном",
+    "Ветеринар",
+    "Фармацевт",
+    "Хирург",
+    "Оператор колл-центра",
+    "Инженер",
+    "Руководитель отдела",
+    "Координатор проектов",
+    "Механик",
+    "Шеф-повар",
+    "Бариста",
+    "Официант",
+    "Сомелье",
+    "Флорист",
+    "Парикмахер",
+    "Косметолог",
+    "Модель",
+    "Фотограф",
+    "Повар",
+    "Преподаватель",
+    "Тату-мастер",
+    "Аниматор",
+    "Гувернантка",
+    "Воспитатель",
+    "Няня",
+    "Риелтор",
+    "Консультант",
+    "Оператор станков",
+    "Инструктор",
+    "Журналист",
+    "Редактор",
+    "Артист",
+    "Актёр",
+    "Продюсер",
+    "Режиссёр",
+    "Швея",
+    "Плотник",
+    "Строитель",
+    "Маляр",
+    "Сварщик",
+    "Кладовщик",
+    "Технолог",
+    "Ремонтник",
+    "Промоутер",
+    "Кассир",
+    "Эколог",
+    "Лаборант",
+    "Менеджер по работе с ключевыми клиентами",
+    "Менеджер по логистике",
+    "Пекарь",
+    "Токарь",
+    "Специалист по защите данных",
+    "Охранник",
+    "Монтажник",
+    "Диспетчер",
+    "Ювелир",
+    "Software Engineer",
+    "Web Developer",
+    "Project Manager",
+    "System Administrator",
+    "Graphic Designer",
+    "Marketing Specialist",
+    "Content Manager",
+    "Copywriter",
+    "Sales Specialist",
+    "Financial Analyst",
+    "Accountant",
+    "Customer Service Specialist",
+    "Lawyer",
+    "Secretary",
+    "Translator",
+    "SEO Specialist",
+    "Video Editor",
+    "Doctor",
+    "Nurse",
+    "Teacher",
+    "Educator",
+    "Architect",
+    "Chemist",
+    "Physicist",
+    "Biologist",
+    "HR Specialist",
+    "Logistician",
+    "Economist",
+    "Sales Consultant",
+    "Masseur",
+    "Dentist",
+    "Psychologist",
+    "PPC Specialist",
+    "Website Administrator",
+    "Digital Marketer",
+    "Data Analyst",
+    "Auditor",
+    "QA Tester",
+    "Driver",
+    "Electrician",
+    "HR Manager",
+    "Director",
+    "Coach",
+    "Interior Designer",
+    "Recruiter",
+    "Procurement Manager",
+    "Courier",
+    "Agronomist",
+    "Veterinarian",
+    "Pharmacist",
+    "Surgeon",
+    "Call Center Operator",
+    "Engineer",
+    "Department Head",
+    "Project Coordinator",
+    "Mechanic",
+    "Chef",
+    "Barista",
+    "Waiter",
+    "Sommelier",
+    "Florist",
+    "Hairdresser",
+    "Cosmetologist",
+    "Model",
+    "Photographer",
+    "Cook",
+    "Lecturer",
+    "Tattoo Artist",
+    "Animator",
+    "Governess",
+    "Kindergarten Teacher",
+    "Nanny",
+    "Realtor",
+    "Consultant",
+    "Machine Operator",
+    "Instructor",
+    "Journalist",
+    "Editor",
+    "Performer",
+    "Actor",
+    "Producer",
+    "Director",
+    "Seamstress",
+    "Carpenter",
+    "Builder",
+    "Painter",
+    "Welder",
+    "Storekeeper",
+    "Technologist",
+    "Repairman",
+    "Promoter",
+    "Cashier",
+    "Ecologist",
+    "Lab Technician",
+    "Key Account Manager",
+    "Logistics Manager",
+    "Baker",
+    "Turner",
+    "Data Protection Specialist",
+    "Security Guard",
+    "Installer",
+    "Dispatcher",
+    "Jeweler",
+    "Front-End Developer",
+    "Back-End Developer",
+    "Full-Stack Developer",
+    "JavaScript Developer",
+    "React Developer",
+    "Node.js Developer",
+    "Angular Developer",
+    "Vue.js Developer",
+    "Python Developer",
+    "Java Developer",
+    "Ruby Developer",
+    "PHP Developer",
+    "Go Developer",
+    "C# Developer",
+    "C++ Developer",
+    "Swift Developer",
+    "Kotlin Developer",
+    "Android Developer",
+    "iOS Developer",
+    "Mobile App Developer",
+    "Blockchain Developer",
+    "DevOps Engineer",
+    "Cloud Architect",
+    "Database Administrator",
+    "Data Scientist",
+    "Machine Learning Engineer",
+    "AI Specialist",
+    "Cybersecurity Specialist",
+    "Network Administrator",
+    "IT Support Specialist",
+    "Game Developer",
+    "Unity Developer",
+    "Unreal Engine Developer",
+    "Scrum Master",
+    "Product Owner",
+    "Agile Coach",
+    "Software Architect",
+    "QA Engineer",
+    "Technical Writer",
+    "Business Analyst",
+    "Technical Support Engineer"
+  ];
+
